+++ server/controllers/questions.js
@@ -4,6 +4,10 @@ const questionsRouter = require('express').Router()
 questionsRouter.put('/:questionId', async (req, res) => {
   console.log("put request to update a question")
   const questionId = Number(req.params.questionId)
+  if (isNaN(questionId)) {
+    res.status(400).end()
+    return;
+  }
   try {
     const result = await pool.query("UPDATE question SET text=$1 WHERE question_id=$2", [req.body.text, questionId])
     if (result.rowCount > 0) {
@@ -19,6 +23,10 @@ questionsRouter.put('/:questionId', async (req, res) => {
 
 questionsRouter.delete('/:questionId', async (req, res) => {
   const questionId = Number(req.params.questionId)
+  if (isNaN(questionId)) {
+    res.status(400).end()
+    return;
+  }
   try {
     const result = await pool.query("DELETE FROM question WHERE question_id=$1", [questionId])
     if (result.rowCount > 0) {
@@ -34,6 +42,10 @@ questionsRouter.delete('/:questionId', async (req, res) => {
 questionsRouter.get('/:questionId/answeroptions', async (req, res) => {
   console.log("received get request for answer options of a question")
   const questionId = Number(req.params.questionId)
+  if (isNaN(questionId)) {
+    res.status(400).end()
+    return;
+  }
   try {
     const result = await pool.query('SELECT * FROM answer_option WHERE question_id=$1', [questionId])
     if (result.rowCount > 0) {
@@ -51,6 +63,10 @@ questionsRouter.get('/:questionId/answeroptions', async (req, res) => {
 questionsRouter.post('/:questionId/answeroptions', async (req, res) => {
   console.log('add answer option');
   const questionId = Number(req.params.questionId)
+  if (isNaN(questionId)) {
+    res.status(400).end()
+    return;
+  }
   const query = {
     text: 'INSERT INTO answer_option (question_id, text, correct) VALUES ($1,$2,$3) RETURNING answer_option_id',
     values: [questionId, req.body.text, req.body.correct],
